feat: Include "intermediate" episodes
The proposed changes query the episode API one single time to retrieve
the whole episodes range.
The main purpose of this is to support series with "intermediate"
episodes (e.g. 2.5, between 2 and 3), that were previously ignored by
the algorithm. The final effect was to actually download them in place
of integer episodes, since they were always the last element in the
selected range (which was instead assumed to be one-element long).

An example of such a series is "Usagi Drop".

https://www.animeunity.so/anime/305-usagi-drop